requirements.txt
certifi==2023.7.22
contourpy==1.1.0
datasets==2.14.5
huggingface-hub==0.16.4
idna==3.4
librosa==0.10.1
matplotlib==3.7.2
numpy==1.24.4
openai-whisper==20230918
pandas==2.1.0
PyYAML==6.0.1
requests==2.31.0
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
soundfile==0.12.1
soxr==0.3.6
tokenizers==0.13.3
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.66.1
transformers==4.33.1
jiwer
